change count type to floate in DcdMapProvider

+++ roveranalyzer/simulators/opp/provider/hdf/DcDGlobalPosition.py
@@ -21,7 +21,7 @@ class DcdGlobalMapKey:
         Y: float,
     }
     types_global_raw_csv_col = {
-        COUNT: int,
+        COUNT: float,
         NODE_ID: str,  # contains comma separated list will later be parsed to int list
     }
